blogpost/requirements.txt
Django==4.2.5
django_ckeditor==6.7.0
